fix tests failings + reviewer's suggestion applied

+++ modules/extension/tile-client/src/main/java/org/geotools/tile/TileService.java
@@ -118,7 +118,8 @@ public abstract class TileService {
      * @param scaleFactor Scale-factor (0-100)
      * @return Zoom-level
      */
-    public int getZoomLevelFromMapScale(ScaleZoomLevelMatcher zoomLevelMatcher, int scaleFactor) {
+    public int getZoomLevelFromMapScale(
+            ScaleZoomLevelMatcher zoomLevelMatcher, double scaleFactor) {
         // fallback scale-list
         double[] scaleList = getScaleList();
         assert (scaleList != null && scaleList.length > 0);
@@ -155,7 +156,7 @@ public abstract class TileService {
      *     selected
      */
     public int getZoomLevelToUse(
-            ScaleZoomLevelMatcher zoomLevelMatcher, int scaleFactor, boolean useRecommended) {
+            ScaleZoomLevelMatcher zoomLevelMatcher, double scaleFactor, boolean useRecommended) {
         if (useRecommended) {
             return getZoomLevelFromMapScale(zoomLevelMatcher, scaleFactor);
         }
@@ -202,7 +203,7 @@ public abstract class TileService {
 
     public Set<Tile> findTilesInExtent(
             ReferencedEnvelope _mapExtent,
-            int scaleFactor,
+            double scaleFactor,
             boolean recommendedZoomLevel,
             int maxNumberOfTiles) {
